백준 사칙연산 5(a/b)2

+++ sc/src/number1/Main.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int a = scanner.nextInt(), b=scanner.nextInt();
+        double a = scanner.nextLong(), b=scanner.nextLong();
         System.out.println(a/b);
     }
 }
